+++ android/src/main/java/com/bitmovin/player/reactnative/Types.kt
@@ -18,6 +18,7 @@ enum class UserInterfaceType {
 data class PictureInPictureConfig(
     val isEnabled: Boolean = false,
     val shouldEnterOnBackground: Boolean = false,
+    val shouldExitOnForeground: Boolean = false,
 )
 
 enum class PictureInPictureAction {