+++ android/src/main/java/com/bitmovin/player/reactnative/converter/JsonConverter.kt
@@ -728,6 +728,7 @@ fun Thumbnail.toJson(): Map<String, Any> = mapOf(
 fun Map<String, Any?>.toPictureInPictureConfig(): PictureInPictureConfig = PictureInPictureConfig(
     isEnabled = getBooleanOrNull("isEnabled") ?: false,
     shouldEnterOnBackground = getBooleanOrNull("shouldEnterOnBackground") ?: false,
+    shouldExitOnForeground = getBooleanOrNull("shouldExitOnForeground") ?: false,
 )
 
 fun List<String>.toPictureInPictureActions(): List<PictureInPictureAction> = mapNotNull {